+++ crates/core/src/services.rs
@@ -6,5 +6,5 @@ pub mod user_info;
 pub use core::CoreServices;
 
 pub use repository::{Repository, RepositoryService, Transaction, read_only_transaction, transaction};
-pub use user::UserService;
-pub use user_info::UserInfoService;
+pub use user::UserRepository;
+pub use user_info::UserInfoRepository;